requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Data Processing
pandas==2.2.3
pdfplumber==0.11.4
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5

# AI
anthropic==0.40.0

# Plaid (bank/credit card connections)
plaid-python==29.0.0

# Market Data
yfinance==0.2.48

# Encryption
cryptography==44.0.0

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
